// src/interfaces/response-body/product-purchase.ts
export interface ProductPurchaseResponse{
    id: number;
    cat_id: string;
    sku: string;
    qty_round_option: string;
    model: string;
    name: string;
    vendor_name: string;
    selling_unit: string;
    upc_code: string;
    suggested_retail: string;
    your_cost: string;
    image: string;
    shelf_pack: string;
    velocity_code: string;
    total_rec_count: string;
    program_number: string;
    quantity: number;
    item_price: number;
}